ssdrotated_lld/requirements.txt
backcall==0.1.0
cffi==1.11.5
cycler==0.10.0
Cython==0.28.5
decorator==4.3.0
ipython==5.8.0
ipython-genutils==0.2.0
jedi==0.12.1
kiwisolver==1.0.1
line-profiler==2.1.2
matplotlib==2.2.3
numpy==1.15.0
opencv-python==3.4.2.17
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
torch==0.4.0
torchvision==0.2.1
traitlets==4.3.2
wcwidth==0.1.7
